+++ README.md
@@ -8,3 +8,6 @@ Currently
 
 1. stores in memory the available products
 1. displays the available products
+1. Creates a customer order for random selection of products
+1. Bakes the products in the order
+1. Packages the order
