+++ service/bakers/cake_baker_service.go
@@ -1,4 +1,4 @@
-package service
+package bakers
 
 import (
 	"fmt"
