+++ service/products/product_service.go
@@ -1,4 +1,4 @@
-package service
+package products
 
 import (
 	"fmt"
