+++ service/products/product_service_test.go
@@ -1,4 +1,4 @@
-package service
+package products
 
 import (
 	"errors"
